Move all logic into '__init__.py'.
Leave behind BBB shims in 'meld3/meld3.py'.
